Add files via upload
This is a machine learning model for classifying SMS messages as either spam or ham (not spam). The model was trained on a dataset of SMS messages labeled as spam or ham.

The model uses natural language processing techniques to extract features from the text of each message, and then applies a machine learning algorithm to classify the message as spam or ham.

This repository includes the code for training and testing the model, as well as a pre-trained model that can be used for classification of new SMS messages.